requirements.txt
numpy==1.21.4
matplotlib==3.5.0
pandas==1.3.4
requests~=2.28.2
python-binance